Move getPhrases to characters folder
